+++ src/main/java/firok/mtim/block/BlockSaplings.java
@@ -1,0 +1,11 @@
+package firok.mtim.block;
+
+import net.minecraft.block.BlockSapling;
+
+public class BlockSaplings extends BlockSapling
+{
+	public BlockSaplings()
+	{
+		super();
+	}
+}