+++ src/main/java/firok/mtim/block/BlockWood.java
@@ -1,0 +1,8 @@
+package firok.mtim.block;
+
+import net.minecraft.block.BlockLog;
+
+public class BlockWood extends BlockLog
+{
+	;
+}