+++ src/main/java/firok/mtim/util/Keys.java
@@ -4,12 +4,17 @@ import firok.mtim.MoreTinkersMaterials;
 
 public interface Keys
 {
+	// 前缀
 	String prefMolten="molten_";
 	String prefBlock="block_";
 	String prefOre="ore_";
 	String prefIngot="ingot_";
+	String prefLog="log_";
+	String prefLeaf="leaf_";
+	String prefSapling="sapling_";
 	String prefRL= MoreTinkersMaterials.MOD_ID+':';
 
+	// 材料
 	String nameRoyalAlloy="royal_alloy";
 	String nameStellarium ="stellarium";
 	String nameInkPowder="ink_powder";
@@ -31,6 +36,13 @@ public interface Keys
 	String nameWitherium="witherium";
 	String nameStellariumObsidian="stellarium_obsidian";
 
+	// 植物
+	String nameHura="hura";
+	String logHura=prefLog+nameHura;
+	String leafHura=prefLeaf+nameHura;
+	String saplingHura=prefSapling+nameHura;
+
+	// 液体
 	String moltenRoyalAlloy=prefMolten+nameRoyalAlloy;
 	String moltenStellarium=prefMolten+nameStellarium;
 	String moltenCinnabar=prefMolten+nameCinnabar;
@@ -40,6 +52,7 @@ public interface Keys
 	String moltenInertWitherium=prefMolten+nameInertWitherium;
 	String moltenWitherium=prefMolten+nameWitherium;
 
+	// 矿块
 	String blockStellarium=prefBlock+ nameStellarium;
 	String blockInkPowder=prefBlock+nameInkPowder;
 	String blockCinnabar=prefBlock+nameCinnabar;
@@ -51,6 +64,7 @@ public interface Keys
 	String blockWitherium=prefBlock+nameWitherium;
 	String blockStellariumObsidian=prefBlock+nameStellariumObsidian;
 
+	// 原矿
 	String oreCinnabar=prefOre+nameCinnabar;
 	String oreInkPowder=prefOre+nameInkPowder;
 	String oreImmersedSilver=prefOre+nameImmersedSilver;
@@ -60,6 +74,7 @@ public interface Keys
 	String oreInertWitherium=prefOre+nameInertWitherium;
 	String oreWitherium=prefOre+nameWitherium;
 
+	// 矿锭
 	String ingotStellarium=prefIngot+nameStellarium;
 	String ingotRoyalAlloy=prefIngot+nameRoyalAlloy;
 	String ingotImmersedSilver=prefIngot+nameImmersedSilver;
@@ -68,6 +83,7 @@ public interface Keys
 	String ingotInertWitherium=prefIngot+nameInertWitherium;
 	String ingotWitherium=prefIngot+nameWitherium;
 
+	// 材料颜色
 	int colorStellarium=Colors.Tomato;
 	int colorRoyalAlloy=Colors.Yellow;
 	int colorSpiderLeg =Colors.IndianRed;
@@ -83,11 +99,13 @@ public interface Keys
 	int colorInertWitherium=Colors.DarkGray;
 	int colorWitherium=Colors.DimGray;
 
+	// 属性
 	String nameTraitIcy="icy";
 	String nameTraitLuxurious="luxurious";
 	String nameTraitRadiant="radiant";
 	String nameTraitSwitching="switching";
 
+	// 属性颜色
 	int colorTraitIcy=Colors.Snow;
 	int colorTraitLuxurious=Colors.Gold;
 	int colorTraitRadiant=Colors.Coral;