+++ changelog.md
@@ -1,9 +1,15 @@
 # 变更日志 Changelog
 
+## 0.1.8 _2019-11-24 20:00_
+
+* 增加方块 added blocks
+  * 胡拉木原木 hura log block
+  * 胡拉木树苗 hura sapling block
+
 ## 0.1.7 _2019-11-23 22:00_
 
 * 增加方块 added blocks
-  * 恒星黑曜石 stellarium obsidian
+  * 恒星黑曜石 stellarium obsidian block
 
 ## 0.1.6 _2019-11-22 19:00_